fix: address conflict with configurations
The previous configuration allowed for `allow-hosts` to be passed
without `--disable-socket` to implicitly activate blocking.

This was confusing and inconsistent so we now require it to enable the
global behavior.

This introduces a little more complexity to the setup method, but is
more clear as to what is actually happening.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>