test: correct execution runner for coverage report
Prior refactoring to use `poetry` for dependency management, the
`coverage` command was executed directly. This was changed to use
`poetry run` to ensure the correct version of `coverage` was used.

However, the `coverage report` command was not updated to use `poetry
run` and thus was not using the correct version of `coverage`.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>